Authenticate /tasks endpoint
Not sure I like how it turned out now. The form now uses multipart for
data, which makes the authorization work in Swagger UI (/docs) but makes
it more complicated to use as pure API. It also forces to use `username`
field to identify a user, not sure how to change it to `email`. That
dictates the structure of API auth endpoints.

I guess that's a trade off I need to make between working /docs
and beauty of endpoint interface.